Extract VersionedFilesService::Object
